styling and error solving

+++ passop/src/components/Manager.jsx
@@ -48,21 +48,54 @@ const Manager = () => {
     };
 
     const savePassword = () => {
+        if(form.site.length >3 && form.username.length > 3 && form.password.length> 3){
+
         setPasswordArray([...passwordArray,{...form,id: uuidv4()}])
         localStorage.setItem("passwords", JSON.stringify([...passwordArray,{...form,id: uuidv4()}]))
         console.log([...passwordArray, form])
+        setForm({site:"", username:"",password:""})
+        toast('Save Succesfully', {
+            position: "top-center",   //For Toast
+            autoClose: 5000,
+            hideProgressBar: false,
+            closeOnClick: true,
+            pauseOnHover: true,
+            draggable: true,
+            progress: undefined,
+            theme: "light",
+        });
+    }
+    else{
+        toast('Err : PAssword not saved')
+    }
     };
 
     const deletePassword = (id)=>{
+        let c = confirm("Do you Really wanted to delete")
+        if(c){
         const updatedArray = passwordArray.filter(item => item.id != id)
         setPasswordArray(updatedArray)
-        localStorage.setItem("passwords",JSON.stringify(passwordArray.filter(item => item.id != id)))
+        localStorage.setItem("passwords",JSON.stringify(passwordArray.filter(item => item.id != id)));
+        toast('Password Deleted Succesfully', {
+            position: "top-center",   // Toast
+            autoClose: 5000,
+            hideProgressBar: false,
+            closeOnClick: true,
+            pauseOnHover: true,
+            draggable: true,
+            progress: undefined,
+            theme: "light",
+        })
         
+        }
     }
 
     const editPassword = (id)=>{
         console.log("button is clicked",id)
         setForm(passwordArray.filter(i=>i.id === id)[0])   // for populating the details of the inputs
+        setPasswordArray(passwordArray.filter(item => item.id !=id)) //jab hum phle bina is statement ke edit krte the to ek new set add hota tha poorana vaal 
+                                                                    //  bhi rhta tha isse poorana vaala delete ho jaayega or new edited array aa jayega
+        console.log("half work done")
     }
 
 
@@ -94,7 +127,7 @@ const Manager = () => {
             <ToastContainer />
 
 
-            <div className="mycontainer">
+            <div className="p-2 md:p-0 md:mycontainer min-h-[81.3vh]">
                 <h1 className='text-4xl font-bold text-center'><span className='text-green-800'>&lt;</span>Pass
                     <span className='text-green-800 text-4xl'>Op/</span><span className='text-green-800'>&gt;</span>
                 </h1>
@@ -108,8 +141,9 @@ const Manager = () => {
                         className="rounded-full border border-green-700 w-full p-4 py-1"
                         type="text"
                         name='site'
+                        id='site'
                     />
-                    <div className='flex w-full gap-8 justify-between' id='cc'>
+                    <div className='flex flex-col md:flex-row w-full gap-8 justify-between' id='cc'>
                         <input
                             value={form.username || ''}
                             onChange={handleChange}
@@ -117,6 +151,7 @@ const Manager = () => {
                             placeholder='Enter Website username'
                             className="rounded-full border border-green-700 w-full p-4 py-1"
                             name='username'
+                            id='username'
                         />
 
                         <div className="relative">
@@ -148,7 +183,7 @@ const Manager = () => {
                 </div>
                 <div className="passwords"><h2 className='font-bold text-2xl py-4 text-center'>Your passwords</h2>
                     {passwordArray.length === 0 ? <div>No password to show</div> :
-                        <table className="table-auto  w-full rounded-md overflow-hidden">
+                        <table className="table-auto  w-full rounded-md overflow-hidden mb-10">
                             <thead className='bg-green-800 text-white'>
                                 <tr>
                                     <th className='py-2'>Site</th>
@@ -160,7 +195,7 @@ const Manager = () => {
                             <tbody className='bg-green-100'>
                                 {passwordArray.map((item, index) => {
                                     return <tr key={index}>
-                                        <td className='text-center py-2 border border-white flex justify-center items-center ' ><a href={item.site} target='blank'>{item.site}</a>
+                                        <td className='text-center py-2 flex justify-center items-center ' ><a href={item.site} target='blank'>{item.site}</a>
                                             <div className='lordicon size-7 cursor-pointer' onClick={() => { copyText(item.site) }}>
                                                 <lord-icon
                                                     style={{ "width": "20px", "height": "25px", "paddingTop": "6px", "paddingLeft": "2px" }}
@@ -170,7 +205,7 @@ const Manager = () => {
                                             </div>
                                         </td>
 
-                                        <td className='justify-center text-center  py-2 border border-white'>
+                                        <td className=' text-center  py-2 '>
                                             <div className='flex justify-center items-center'>
                                                 <span>{item.username}</span>
                                                 <div className='lordicon size-7 cursor-pointer' onClick={() => { copyText(item.username) }}>
@@ -183,7 +218,7 @@ const Manager = () => {
                                             </div>
 
                                         </td>
-                                        <td className='text-center w-32 py-2 border border-white'>
+                                        <td className='text-center w-32 py-2'>
                                             <div className='flex justify-center items-center'>
                                                 <span>{item.password}</span>
                                                 <div className='lordicon size-7 cursor-pointer' onClick={() => { copyText(item.password) }}>
@@ -196,7 +231,7 @@ const Manager = () => {
                                                 </div>
                                             </div>
                                         </td>
-                                        <td className='text-center w-32 py-2 border border-white'>
+                                        <td className='text-center w-32 py-2'>
                                             <span className='cursor-pointer  mx-1' onClick={()=>{deletePassword(item.id)}}>
                                                 <lord-icon
                                                 src="https://cdn.lordicon.com/skkahier.json"
